Advanced diploma

See merge request maksim_zhitkov/python_advanced_diploma!10

+++ app/crud/image.py
@@ -1,5 +1,5 @@
 """Module for database query operations with the Image model"""
-from typing import List
+from typing import List, Dict
 
 from models.model import Image
 from schemas.tweet_schema import AddTweetSchema
@@ -10,12 +10,19 @@ from sqlalchemy.ext.asyncio import AsyncSession
 async def transform_image_id_in_image_url(
     session: AsyncSession,
     tweet_in: AddTweetSchema,
-):
-    """We change image IDs to URLs to store a list of
-    URLs rather than a list of IDs, which will make it
-    easier to retrieve tweets."""
-    unpack = tweet_in.model_dump()
-    media = await get_image_url(session, unpack["tweet_media_ids"])
+) -> Dict[str, List[str]]:
+    """
+    Функция получает список имен картинок из списка id картинок.
+
+    :param session: Сессия для работы с бд.
+    :param tweet_in: Пришедшие с frontend данные о твите.
+    :return image_ids: Возвращаем данные где вместо списка id в tweet_media_ids будет
+    список имен картинок твита.
+    """
+    unpack: dict = tweet_in.model_dump()
+    media: ScalarResult[str] = await get_image_url(session, unpack["tweet_media_ids"])
+
+    # Удаляем данные о картинке из бд, так как они нам больше не нужны.
     await remove_image_id_in_db(session, unpack["tweet_media_ids"])
     unpack["tweet_media_ids"] = list(media)
     return unpack
@@ -24,8 +31,13 @@ async def transform_image_id_in_image_url(
 async def get_image_url(
     session: AsyncSession,
     image_id_list: List[int],
-) -> ScalarResult[int]:
-    """Function return URL from ID."""
+) -> ScalarResult[str]:
+    """
+    Функция получает список имен картинок.
+
+    :param session: Сессия для работы с бд.
+    :param image_id_list: Список id картинок.
+    :return image_url: Список имен картинок."""
     stmt = select(Image.url).where(Image.id.in_(image_id_list))
     return await session.scalars(stmt)
 
@@ -34,9 +46,16 @@ async def remove_image_id_in_db(
     session: AsyncSession,
     image_id_list: List[int],
 ) -> None:
-    """Removes unnecessary photos from the table"""
+    """
+    Функция удаляет данные о картинке из бд.
+
+    :param session: Сессия для работы с бд.
+    :param image_id_list: Список id картинок.
+    :return None: Ничего не возвращает.
+    """
     for img_id in image_id_list:
         stmt = select(Image).where(Image.id == img_id)
-        img = await session.scalar(stmt)
-        await session.delete(img)
+        img: Image | None = await session.scalar(stmt)
+        if img:
+            await session.delete(img)
     await session.commit()
